+++ docs/examples/stream/stream_example.c
@@ -88,7 +88,7 @@ main( int argc, char **argv ) {
 
 
   // destroying all remaining resources
-  stumpless_destroy_entry( basic_entry );
+  stumpless_destroy_entry_and_contents( basic_entry );
 
 
   return EXIT_SUCCESS;