+++ test/function/config/network_unsupported.cpp
@@ -44,7 +44,7 @@ namespace {
     EXPECT_LT( result, 0 );
     EXPECT_ERROR_ID_EQ( STUMPLESS_TARGET_UNSUPPORTED );
 
-    stumpless_destroy_entry( entry );
+    stumpless_destroy_entry_and_contents( entry );
 
     target->type = STUMPLESS_BUFFER_TARGET;
     stumpless_close_buffer_target( target );